Move functions into new html and file_processins modules
